server/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
jinja2>=3.1.4
aiofiles>=24.1.0

# Database
aiosqlite>=0.20.0

# HTTP
httpx>=0.28.1

# Browser Automation
browser-use==0.12.6
playwright>=1.49.0

# LLM
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0

# Scheduling
apscheduler>=3.10.4

# Utilities
pydantic>=2.10.0
rich>=13.9.4
python-multipart>=0.0.12
